requirements.txt
requests
pandas
numpy
python-dateutil
urllib3
python-dotenv
